move the logic into reporting-page
