+++ plugins/CoreHome/angularjs/reporting-page/reportingpage.directive.html
@@ -4,6 +4,8 @@
 
     <div ng-show="hasNoPage">{{ 'CoreHome_NoSuchPage'|translate }}</div>
 
+    <div class="alert alert-info" ng-show="hasOnlyRawData" ng-bind-html="rawDataMessage"></div>
+
     <div class="row" ng-repeat="widget in pageModel.widgets">
         <div class="col s12 fullWidgetColumn" ng-if="!widget.group" piwik-widget="widget"></div>
